Intégration de l'API à l'enregistrement d'un trajet

+++ modeles/trajet.dao.php
@@ -233,7 +233,7 @@ class TrajetDao{
      * @param integer|null $numero_lieu_arrivee
      * @return void
      */
-    public function insert(?string $heureDep = null,?string $heureArr = null,?int $prix = null,?string $dateDep = null,?int $nbPlace = null,?int $numero_conducteur = null,?int $numero_lieu_depart = null,?int $numero_lieu_arrivee = null): void
+    public function insert(?string $heureDep = null,?string $heureArr = null,?float $prix = null,?string $dateDep = null,?int $nbPlace = null,?int $numero_conducteur = null,?int $numero_lieu_depart = null,?int $numero_lieu_arrivee = null): void
     {
         $query = $this->PDO->prepare("INSERT INTO TRAJET(heureDep, heureArr, prix, dateDep, nbPlace, numero_conducteur, numero_lieu_depart, numero_lieu_arrivee) VALUES (:heureDep, :heureArr, :prix, :dateDep, :nbPlace, :numero_conducteur, :numero_lieu_depart, :numero_lieu_arrivee)");
